+++ src/api/handy/getStrokeSpeedAndDistance.ts
@@ -1,0 +1,31 @@
+// The fastest speed the handy can operate at.
+export const maxStrokeSpeed = 400;
+
+// Max stroke is around 200mm including buffers at the end. Varies +-5mm depending on FW and batch #.
+export const maxStrokeLength = 195;
+
+// The fastest beats per second the handy can operate at with the maximum stroke length.
+export const fastestBps = maxStrokeSpeed / maxStrokeLength;
+
+/**
+ * Calculates the stroke speed and distance from the specified beats per second.
+ */
+export default function getStrokeSpeedAndDistance(bps: number) {
+  let speed = maxStrokeSpeed;
+  let stroke = maxStrokeLength;
+
+  // Decrease stroke speed to decrease BPS
+  if (bps < fastestBps) {
+    speed = maxStrokeLength * bps;
+  }
+
+  // Shorten stroke distance to increase BPS
+  if (bps > fastestBps) {
+    stroke = maxStrokeSpeed / (bps * fastestBps);
+  }
+
+  return {
+    speed,
+    stroke,
+  };
+}