+++ src/components/atoms/HomeButton/HomeButton.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import IconButton from "@material-ui/core/IconButton";
 import HomeIcon from "@material-ui/icons/Home";
 import { useHistory } from "react-router-dom";
+import { Tooltip } from "@material-ui/core";
 
 export default function HomeButton() {
   const history = useHistory();
@@ -11,8 +12,10 @@ export default function HomeButton() {
   };
 
   return (
-    <IconButton color="inherit" onClick={handleClick}>
-      <HomeIcon />
-    </IconButton>
+    <Tooltip title="Home" placement="bottom">
+      <IconButton color="inherit" onClick={handleClick}>
+        <HomeIcon />
+      </IconButton>
+    </Tooltip>
   );
 }