+++ src/components/molecules/ConnectHandy/ConnectHandy.stories.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+import ConnectHandy from "./ConnectHandy";
+
+export default {
+  title: "molecules/ConnectHandy",
+  component: ConnectHandy,
+};
+
+export const Standard = () => <ConnectHandy />;