+++ src/containers/HUD/HUDButtons.tsx
@@ -8,10 +8,10 @@ import SkipNextButton from "components/molecules/buttons/SkipNextButton";
 import SkipPreviousButton from "components/molecules/buttons/SkipPreviousButton";
 import StarButton from "components/molecules/buttons/StarButton";
 import { selectGame } from "common/store/currentGame";
+import ConnectHandy from "components/molecules/ConnectHandy";
 
 const HUDButtons = React.memo(function HUDButtons() {
   const currentGame = useSelector(selectGame);
-
   return (
     <div
       style={{
@@ -36,6 +36,7 @@ const HUDButtons = React.memo(function HUDButtons() {
           starred={currentGame.starred}
         />
       )}
+      <ConnectHandy variant="icon" />
     </div>
   );
 });