+++ public/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented here.
 
 The overall project roadmap and progress can be viewed on our [Trello Board](https://trello.com/b/JAIcnXRf/fap-instructor)
 
+## 2021-07-14
+
+- Added Handy hardware support!
+
 ## 2021-06-23
 
 - Added button to view previous image